requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-openai
pymupdf
python-dotenv